Add tests and refactor for tests.
git-svn-id: https://geobeagle.googlecode.com/svn@597 eab1e41a-c3b7-11dd-b16b-519e865567fa

+++ branches/sng/GeoBeagle/src/com/google/code/geobeagle/Geocache.java
@@ -28,17 +28,17 @@ import android.os.Parcelable;
  */
 public class Geocache implements Parcelable {
     static interface AttributeFormatter {
-        CharSequence formatAttributes(float difficulty, float terrain);
+        CharSequence formatAttributes(int difficulty, int terrain);
     }
 
     static class AttributeFormatterImpl implements AttributeFormatter {
-        public CharSequence formatAttributes(float difficulty, float terrain) {
+        public CharSequence formatAttributes(int difficulty, int terrain) {
             return (difficulty / 2.0) + " / " + (terrain / 2.0);
         }
     }
 
     static class AttributeFormatterNull implements AttributeFormatter {
-        public CharSequence formatAttributes(float difficulty, float terrain) {
+        public CharSequence formatAttributes(int difficulty, int terrain) {
             return "";
         }
     }
